исправление user.model

+++ artele/users/models.py
@@ -28,8 +28,8 @@ class User(AbstractUser):
                                  null=True,
                                  default='Пользователь',
                                  verbose_name='Фамилия')
-    telegram_chat_id = models.IntegerField(unique=True,
-                                           verbose_name='Телеграмм id')
+    telegram_chat_id = models.BigIntegerField(unique=True,
+                                              verbose_name='Телеграмм id')
     role = models.CharField(max_length=16,
                             choices=Role.choices,
                             default=Role.GUEST,
